+++ main.py
@@ -1,23 +1,25 @@
 # Joseph Woodward
 # CS 30 Q1 AM
-# Oct 13 2020
-# Finals Game (For map assignment)
+# Oct 23 2020
+# Finals Game
 
 # Imports
 import first_day as first
 import second_day as second
 import time as t
-import characters as c
-import map as m
 
 # Intro message
 print("Hello there!\n")
 t.sleep(1)
 print("You currently have 4 hours to prepare for your finals"
-      " tomorrow at Campbell Colleigate\n")
+      " tomorrow at Campbell Colleigate")
 t.sleep(1)
-print("You can chose how to prepare either by studying like a"
-      " good student or trying to sneak some  extra notes into"
+print("\nThese finals are extra important because your "
+      "application to the U of R is going to be \naccepted or rejected "
+      "depending in your grade, so make sure to prepare well...")
+t.sleep(1)
+print("\nYou can chose how to prepare either by studying like a"
+      " good student or trying to sneak some \nextra notes into"
       " your class....\nBut watch out! If you get caught cheating"
       " it's gonna cost you big time.\n")
 t.sleep(1)
@@ -25,7 +27,7 @@ print("Here's a general idea of what your room looks like:")
 first.print_map1()
 t.sleep(1)
 print("\nIf you look hard enough, you might be able to find something "
-      " useful around here...")
+      "useful around here...")
 print("Type 'map' to check back in on this anytime")
 t.sleep(1)
 print("\nGood luck!")
@@ -48,6 +50,5 @@ print("You enter Campbell Colleigate ready to do your finals...")
 t.sleep(1)
 
 # Second day
-print("Type 'map' to see your map and available classes")
 grade = 0
 second.second_day(grade)