+++ changelog.md
@@ -22,3 +22,22 @@ Prerelease - Oct 15 2020
 Prerelease - Oct 16 2020
 - Added map for first day
 - Made it so that maps update if desk/bookshelf is empty or you've been to a class
+
+Prerelease - Oct 22 2020
+- Added endings
+- Changed intro to align with endings
+- Formated text better so game is easier to read
+
+Alpha - Oct 23 2020
+- Added endings
+- Made lists of things you can study/cheat for update withwhat you've studied/cheated on
+- Removed psuedocode
+- Fixed a syntax error that was causing the game to not run
+
+Beta - Oct 27 2020
+- Ran beta tests
+- Fixed typos
+- Fixed syntax error that was causing english cheating to not owrk
+
+V 1.0 - Oct 27 2020
+- Released the game